requirements-dev.txt
# Development and Code Quality Dependencies
# Install with: pip install -r requirements-dev.txt

# Code Formatting and Linting
black>=23.0.0                    # Uncompromising code formatter
isort>=5.12.0                    # Import sorting
flake8>=6.0.0                    # Linting and style guide enforcement
flake8-docstrings>=1.7.0         # Docstring conventions
flake8-import-order>=0.18.0      # Import order checking
flake8-bugbear>=23.0.0           # Additional bug and design problems

# Type Checking
mypy>=1.0.0                      # Static type checker
types-requests>=2.31.0           # Type stubs for requests
types-python-dateutil>=2.8.0    # Type stubs for dateutil

# Testing and Coverage
pytest>=7.4.0                   # Testing framework
pytest-cov>=4.1.0               # Coverage plugin for pytest
pytest-asyncio>=0.21.0          # Async testing support
pytest-mock>=3.11.0             # Mocking utilities
coverage[toml]>=7.3.0            # Test coverage measurement

# Security and Dependency Scanning
bandit>=1.7.5                   # Security linter
safety>=2.3.0                   # Dependency vulnerability scanner
pip-audit>=2.6.0                # PyPI package vulnerability scanner
detect-secrets>=1.4.0           # Automated secret detection and baseline management

# Pre-commit and Git Hooks
pre-commit>=3.4.0               # Git hook management

# Documentation
sphinx>=7.1.0                   # Documentation generator
sphinx-rtd-theme>=1.3.0         # Read the Docs theme
pydocstyle>=6.3.0               # Docstring style checker

# Development Utilities
ipython>=8.14.0                 # Enhanced Python shell
jupyterlab>=4.0.0               # Notebook environment
memory-profiler>=0.61.0         # Memory usage profiling
line-profiler>=4.1.0            # Line-by-line profiling

# Performance and Monitoring
psutil>=5.9.0                   # System and process utilities
py-spy>=0.3.14                  # Sampling profiler

# API Documentation
fastapi>=0.104.0                # For API documentation generation (optional)
uvicorn>=0.23.0                 # ASGI server (optional)

# Advanced Testing
hypothesis>=6.88.0              # Property-based testing
faker>=19.6.0                   # Fake data generation for testing

# Markdown Documentation Tools
markdown>=3.5.0                 # Markdown parsing/formatting utilities
pyyaml>=6.0.0                   # YAML parser for metadata validation
requests>=2.31.0                # HTTP client for link checking

# MkDocs Documentation Site Generator
mkdocs>=1.5.0                   # Static site generator for documentation
mkdocs-material>=9.5.0          # Material theme for MkDocs
mkdocs-minify-plugin>=0.7.0     # Minify HTML/JS/CSS for production
pymdown-extensions>=10.7.0      # Markdown extensions (Mermaid, syntax highlighting)
